pyproject.toml
[project]
name = "20251207093313"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain>=0.3.0",
    "python-dotenv>=1.0.0",
]
